remove CONST_Tiger_Data_Path in favour of a command line option
The path for the Tiger source data is only needed when importing
the Tiger data and that is usually done only once. So it makes much
more sense to state it on the command line, when using a different
default.

We have never really advertised the option except in the developer
documentation. So it should be fairly safe to remove it without a
deprecation phase.