allow for multiple token analysis modules
Huge refactoring of how the former ICUNameProcessor is called.
The ICU tokenizer configuration now allows to configure multiple
'token analysers'. This is the part that formerly created the
name variants. The analysers are choosen during the name
preprocessing by tagging the name with the 'analyzer' attribute.

As an example we currently define a 'ref' analyzer which does
never produce any variants. In the future other kinds of analyzers
can be added.