switch from search groups to flat search ranking
The concept of search groups does not allow for finer grained ranking
decisions for both the search description and in result ranking.
This removes the search groups in favour of a simple flat rank order
of search descriptions and results. The early break condition now
depends on how far away the rank of the search descriptions is from
the best ranked result found so far.