mark partially covering address lines
Flag lines in the place_addressline table, where the place area of
the term making the address does not completely cover the street.
These lines can then be rechecked on the fly when making addresses
for dependent places, i.e. house numbers and POIs.

This slightly simplifies the code that determines if rechecking is
necessary and also covers cases where there is only one address
term available for an address rank and therefore the need to recheck
is not obvious.