migration: partial with null triggers old-style handling
The partial column cannot be filled during migration because
an appropriate query would take more than 6 hours. Instead use
a null value in partial to fall back to the previous behaviour.
This allows to slowly switch to using the partial column during
updates. Given that place_addressline rows for a place are always
updated at the same time, the query for address lookup will
either use the old style or the new style for all results. So
this should eventually lead to consistent results.