hand full and partial name and place tokens separately into SQL code
The separate token types will be needed later for the search index.
Also switches to only using full tokens for matching up addr:place
places. This is what effectively happened already because we required
all tokens to be present to match. Only using the full tokens means
that the comparison should be slightly faster.